[Refactor] Producer 및 Controller 패키지 변경 및 상수주입식으로 변경 #1
