+++ src/main/java/com/logmate/streaming/producer/config/KafkaProducerConfig.java
@@ -1,8 +1,10 @@
 package com.logmate.streaming.producer.config;
 
+import com.logmate.streaming.common.constant.kafka.KafkaConstant;
 import com.logmate.streaming.common.dto.SpringBootParsedLog;
 import java.util.HashMap;
 import java.util.Map;
+import lombok.RequiredArgsConstructor;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.springframework.context.annotation.Bean;
@@ -13,12 +15,15 @@ import org.springframework.kafka.core.ProducerFactory;
 import org.springframework.kafka.support.serializer.JsonSerializer;
 
 @Configuration
+@RequiredArgsConstructor
 public class KafkaProducerConfig {
+
+  private final KafkaConstant kafkaConstant;
 
   @Bean
   public ProducerFactory<String, SpringBootParsedLog> producerFactory() {
     Map<String, Object> configProps = new HashMap<>();
-    configProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
+    configProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaConstant.server.URL);
     configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
     configProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
     configProps.put(JsonSerializer.ADD_TYPE_INFO_HEADERS, false); // 불필요한 헤더 제거
