+++ src/main/java/com/logmate/streaming/producer/log/producer/KafkaLogProducer.java
@@ -1,7 +1,7 @@
-package com.logmate.streaming.producer.service;
+package com.logmate.streaming.producer.log.producer;
 
+import com.logmate.streaming.common.constant.kafka.KafkaConstant;
 import com.logmate.streaming.common.dto.SpringBootParsedLog;
-import com.logmate.streaming.common.topic.KafkaTopic;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
@@ -12,10 +12,11 @@ import reactor.core.publisher.Mono;
 public class KafkaLogProducer {
 
   private final KafkaTemplate<String, SpringBootParsedLog> kafkaTemplate;
+  private final KafkaConstant kafkaConstant;
 
   public Mono<Void> sendLog(SpringBootParsedLog logMessage) {
     return Mono.fromRunnable(() ->
-        kafkaTemplate.send(KafkaTopic.LOG_TOPIC, logMessage)
+        kafkaTemplate.send(kafkaConstant.topic.LOG_TOPIC, logMessage)
     );
   }
 }
